# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The responses of the open api in swagger focus on the data field rather than the code and the message fields",
        "title": "Rainbow-APP-Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "console.nftrainbow.cn",
    "basePath": "/apps",
    "paths": {
...
        "/poap/activity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get User Activity list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Get User Activity list",
                "operationId": "GetUserActivities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "activityId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractAddress",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActivityQueryResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Add Activity",
                "operationId": "AddActivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "activity config",
                        "name": "activity_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActivityReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/activity/h5": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Set H5 Page Config",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Set H5 Config",
                "operationId": "SetActivityH5Config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "poap_h5_config",
                        "name": "poap_h5_config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.H5Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.H5Config"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/activity/result/{activity_code}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Mint Result list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Get Mint Result list",
                "operationId": "GetMintResultList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "activity_code",
                        "name": "activity_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.POAPResultQueryResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/activity/result/{activity_code}/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Mint Result",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Get Mint Result",
                "operationId": "GetMintResult",
                "parameters": [
                    {
                        "type": "string",
                        "description": "activity_code",
                        "name": "activity_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.POAPResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/activity/{activity_code}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Activity detail info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Get Activity detail",
                "operationId": "GetActivityDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "activity_code",
                        "name": "activity_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Update Activity",
                "operationId": "UpdateActivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "activity_id",
                        "name": "activity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update activity request",
                        "name": "update_activity_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateActivityReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/count/{address}/{activity_code}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Mint Count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Get Mint Count",
                "operationId": "GetMintCount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "activity_code",
                        "name": "activity_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/h5": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mint By H5",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Mint By H5",
                "operationId": "MintByH5",
                "parameters": [
                    {
                        "description": "mint request",
                        "name": "mint_req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.MintReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.POAPResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/poap/wallet/user": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add user wallet profile, so backend can get user phone from anyweb",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POAP"
                ],
                "summary": "Add user wallet profile, so backend can get user phone from anyweb",
                "operationId": "AddWalletUser",
                "parameters": [
                    {
                        "type": "string",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "wallet",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginutils.CommonMessage"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "appService_errors.RainbowAppServiceErrorDetailInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ActivityQueryResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                }
            }
        },
        "models.ActivityReq": {
            "type": "object",
            "required": [
                "activity_type",
                "amount",
                "app_id",
                "description",
                "max_mint_count",
                "name"
            ],
            "properties": {
                "activity_picture_url": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "app_id": {
                    "type": "integer"
                },
                "app_name": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "contract_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "integer",
                    "default": -1
                },
                "is_phone_white_list_opened": {
                    "type": "boolean"
                },
                "max_mint_count": {
                    "type": "integer"
                },
                "metadata_uri": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nft_configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NFTConfig"
                    }
                },
                "start_time": {
                    "type": "integer",
                    "default": -1
                },
                "white_list_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WhiteListInfo"
                    }
                }
            }
        },
...
    }
}